refactor: vector database adapter: remove QdrantVectorAdapter and consolidate types
- Deleted the QdrantVectorAdapter implementation and its associated methods from adapter.py.
- Moved relevant data structures (CollectionSchema, VectorRecord, VectorMatch, TextDocument) to a new types.py file for better organization.
- Updated __init__.py to reflect the removal of the adapter and the addition of types.
- Removed unit tests related to QdrantVectorAdapter from test_adapter.py.
